Make check_references independant of the package
check_references runs on the resource files as they exist at the top
level, not on those within the package. This cannot be done in travis
while using functions from the package in the case where the api is
made available at the package level. If from famplex.api import * is
placed in the packages __init__.py then the package will attempt to
initialize a FamplexGraph when the necessary resources are
unavailable. Removing this single function that is a dependancy on the
package is the smallest change that fixes the situation. Updating
check references to rely on resources within the package will be more
difficult. check_references requires special handling when processing
the resource files that are not needed within the package.